refactor: extract shared hydrateCatalogs to test-utils, revert optimizeDeps
- DRY: move hydrateCatalogs into tests/engine/test-utils.js, used by
  both snapshot.test.js and generate-fixtures.js
- Revert optimizeDeps change (only needed for dev server, not build)
